backend/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
google-genai>=0.1.0
google-generativeai>=0.3.0
chromadb>=0.4.0
Pillow>=10.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
sentence-transformers>=2.2.0
python-dotenv>=1.0.0
gunicorn>=21.0.0